As part of the Nuke integration, we can no longer guarantee that included/required files are in the current directory.  So, we now prepend $GALLERY_BASEDIR to all included/required files.
There's a bug in PHP 4.0.1pl2 that require_once() barfs if you
give it a variable instead of a constant, so we now use require()
instead.  Once upon a time when we had init.php set up as an
auto_prepend, require_once was more valuable for us.  For now,
it's not really necessary.