* Fixed a bug whereby resetting click counts was timing out for large albums.  Reported by Ed.
* Added public picture commenting feature. SF #404094
Specifics:
a.  Album owners can toggle it on/off per album
b.  Anyone can post a new comment if it is toggled on.
c.  People that can write to an album can delete individual comments.
d.  Comments consist of Name, Date, IPNumber, and Text.
e.  An '*' appears next to the captions of any photo with
comments.