Added an argument to getAlbumDirURL that tells it what you want the URL for.  If you specify "highlight" it will give you the local URL. If you specify anything else it'll try to give you a mirror.
Highlights are typically shown for many albums at once, and it's slow
to check each different album just for one image.  They're also
typically pretty small.  So, if the request is for a highlight, don't
mirror it.

Also made getPhoto() more robust.  It checks to see if the index is
out of bounds before returning the value.  Since this method now
returns a reference, if we return a value that's out of bounds it
will get created in the photos array (!)