DatabaseStorage is on its way!
We now support connecting to a remote database using ADOdb (which I haven't
checked in 'cause it's big and I'm not sure I want to check in all the
various support files that come with it).

Implemented the newNucleus() method, which uses a sequence to generate the
next id to use.