Fixed a serious bug where Win32 doesn't create backups of the album data files.
It seems that Windows doesn't like to rename a file to "album.dat.bak" for
some unknown (to me) reason.  It is happy with all kinds of other file names
(like "album.bak.bak").  I've changed it such that on Win32 it calls the file
"album.bak" instead, and that appears to work.