We can now automatically generate a database schema from our class hierarchy!
I tightened up the GalleryClassMetaInfo so that it provides enough
information about each item so that we can readily generate an
appropriate schema.  The DatabaseStorage class can be instructed to
create a new schema (this works), or update the existing one (not yet
finished).  It's set up so that we can readily generate schemas for
lots of different types of databases, basically anything that is
supported by ADOdb.