Reworked the class meta info yet again.
I decided that putting all the meta info into a single class would
make it more difficult for us to easily drop in new types.  Instead,
it's better to have all the class meta info in a method inside the
class itself.

In order to make this as painless as possible, I've added a new set
of javadoc-like tags into each variable definition.  You can specify
the type of the variable at that time, and then I've got a perl script
that will create the accessor functions and the meta-info definition
automatically for you.  The DatabaseStorage strategy then reads this
information to automatically create the database schema.