Major round of changes!
The locking system that I originally envisioned was not working out.
Because of the interconnectedness of the object model, it's
fundamentally impossible to be sure that any local lock is going to
lock all the items that you need to lock.  So, I've done away with
local locks altogether.

As part of this process, I've extended DatabaseStorage with
MySqlDatabaseStorage so that I can use MySql specific locks which are
much more efficient than the system I was using before.  The old
system was very portable and we can revive it if we want to provide a
generic locking system for other platforms when we start supporting
them.

GalleryImageToolkit was renamed to GalleryGraphics (because the length
of the name was pushing the 32 bit char limit that causes some OS'
to have a hard time).

All of the interchangeable components (like GalleryPlatform,
GalleryGraphics, DatabaseStorage) are now serviced via factory classes
instead of direct instantiation.  This simplifies the init code
considerably since now you can just provide configuration information
for the component that you want and let the factory take care of
creating it for you if and when that component is needed (for example,
90% of the time we never need the image toolkit because we're just
viewing images).

Tightened up the pseudo-transaction code so that it bombs if you try
to release a lock while you still have uncommitted data.  This will
help application developers to catch bugs a bit quicker.