Did a pass through the templating system and simplified things.
The translation base code has been abstracted up to the GalleryView
level; this forces us to provide the module name every time we render
a template (which avoids the problem of forgetting to specify a
translation base).

The theme activation code is now called once from main.php so that
we don't have to call it in each view.  For the most part we activate
the theme based on the active item, but a view can request to use the
default site-wide theme instead.