Refactored the way that controllers return their results to make it more uniform.  This required touching every controller and controller test, but now they are all normalized.
Implemented a very easy to use system for tagging the "return to"
link onto "modal" operations.  If you mark a {gallery->link} with
'url_return=true' it will add in the return link, which will get
propagated into every form post automatically for you.  If the controller
returns a 'return' value in its results, the main code will automatically
send you back to the source.  This is great for operations like login,
and the AdminModules configure option which send you to a different part
of the site temporarily, then return you when you're done.