Fixed a bug where thumbnails derived from other, smaller thumbnails would not wind up being larger than the smaller thumbnail.  This is because the thumbnail operation won't generate a new thumbnail that is larger than its source.
To fix it, need to look further back the chain to find a preferred
derivative, resized derivative or original data item and generate the
thumbnail from that, being careful to obey all the rules from the
intervening thumbnails (like cropping, etc).

In the process, I refactored the merge-derivative-operation code so
that it no longer cares about mime types.  We weren't using this
anywhere and it was adding complexity.  If it turns out that we need
the mime time to determine how to merge operations, we can add it back
in.