Addition of GalleryStorage::checkPoint() which will commit pending transactions to the DB as well as anything else that we later decide should be tied in.
Addition of DatabaseStorage::_guaranteeTransaction() and the representative calls to ensure
that whenever we make a call which can change the db, we have a transaction.  This allows us
to call checkPoint at any point in time without having to explicitly re-open a new transaction
inline.

Next up:  Adding checkPoing to the migration module.