My recent shutterfly/photoaccess updates make permission updates in upgrade code which fail without a session/active user; so new install of these modules or upgrade via upgrader (not site admin) failed.
Refactored installer/upgrader so now install modules and upgrade
core/modules have an empty session with an admin user as active user.
Updated GalleryEntity::onLoad with new method to detect if it is in
install/upgrade as $session will now be set.
And show debug output in install modules if a module install fails.