Bad news: a number of things (only embedded G2 as far as i see) are broken in the new rewrite module, didn't notice them in the review. - option forceDirect isn't obeyed in all cases - embedded core.DownloadItem doesn't go to standalone G2 (the test was broken already in the old rewrite module. That's why we didn't notice that the functionality was broken with rewrite 1.1.0.
I've fixed the unit tests, so now we see the test failures. (path info test coverage is too small to cover these issues, still).
But I haven't fixed it yet.
I'll have to change the rewrite module (only the UrlGenerators) anyway to make it work with the new GalleryUrlGenerator class, so this will be part
of my future fix. No need that this is fixed independently.