Fix a bug where if the active theme is incompatible with the current APIs, you get bounced to the upgrader, even if the theme doesn't require upgrading.  This happens in the case where you have your own custom theme and you upgrade the rest of G2 around it.
Fixed by changing GalleryCoreApi::loadPlugin to take an optional
errorOnVersionMismatch parameter which will return an error instead of
trying to send you to the upgrader.  We then handle this with our
fallback theme which allows you to navigate to the SiteAdmin and
switch to a different theme (or at least see what's the
incompatibility with your current theme).