Give print services direct access to the preferred version of the image, and don't send the thumbnail if it has postfilter operations.  This allows users to print photos even if they don't have access to the high-res version of the image, and will avoid sending watermarked thumbnails to the print service.
Downside to this is that if the preferred is watermarked, then the user is going
to get the watermarked version.  No easy way around that though because we don't
have anything else to give to the print service (except for maybe the resized,
but that won't have the same level of quality).