Add support for a new 'text' type parameter for blocks.
Based on some work by donkenjo[1], this allows us to store arbitrary
text values in the parameters for a block.  Arbitrary text values
break our existing neutral data format, so I switched over to using
JSON.  This required a lot of little fiddly work in the Javascript --
now we have to convert quotes to &quot; so that we can use them in
generated JS, eg:

  var x = "some data with &quot; in it";

and then in the JS code we have to convert the &quot; back so that
the JS code has the raw data.  We also have to escape our quote values
carefully.  A little tricky, but it works.

[1]  http://gallery.menalto.com/node/51793