Plugin uninstall now gives you DHTML popup confirm dialog (using YUI) instead of taking you to a new page, so that it doesn't break the nice AJAX-ness of the Plugins page.
We now have a gallery-wide css file at modules/core/data/gallery.css
that is included in every page view.  All application-wide common CSS
should go in this file and it should be scoped as tightly as possible
(eg, to div#gallery) so that it doesn't leak out of the application.

AdminModulesVerifyUninstall code, templates and tests are all gone.