Add support for multiple repositories.  There are three repositories: * Officially released plugins * Experimental plugins * Community plugins
You can load up the indices from all of those at once and then on the
plugin download page you can select which version of the plugin you'd
like to install.  You can upgrade from an older version of the plugin
to a newer one, or add in missing language packs.