Fix blocking bug 1620362.
We calculate the progress bar by summing up every package that needs
to be installed, but then we only update the completed action count
for plugins, not for individual packages. Languages don't get counted,
which means that if you install something that has a lot of languages,
the progress bar seems to make no progress.

Incidentally, we were updating the progress bar every time a package
is installed, and if you're installing 30 language packs this tends to
be very, very chatty for little value. I moved the progress bar update
out so that we update once per plugin and then only update for every
5th file.

Also, guarantee a 30 second time limit for each file operation.