Partial fix for bug 1629483.
I fixed one aspect of this bug by moving the repository data under
g2data/cache and adding a function in lib/support/cache.php to clean
it out. Long running G2 installs will probably accumulate a lot of
packages in the cache over time so letting the user have a way to
flush them (even if it's kind of sucky right now to have to use
lib/support -- automated would be better) is a good idea and that's
definitely easier if it's under g2data/cache.

Bump core module 1.1.28.