Don't advise users to upgrade the core for experimental/commmunity repositories.
Right now if you're tracking the experimental repository, you get
alerted that there's a new version of the core module available, but
the links only take you directly to the released version. I'm not sure
that we want to be directing people towards a new version of the
experimental core anyway. So for now, just don't alert users unless
it's a new versinon in the 'released' repository.