Fix reproducible obsolete data problem when adding items.
The problem happens when you have both the exif and sizelimit options
enabled to work on the original image. One of them will go first, then
the second one will fail because it's working off of the original
$item instance, but the first one has already saved it causing the
serial number to increment.

During the process of finding and fixing this, I noticed several other
ItemAddOption and ItemEditOption instances that were acquiring a write
lock but not refreshing the item which can lead to a race condition
(highly rare for ItemAdds but more probably for ItemEdits). Fixed
those, too.