Fix two minor upgraded related bugs in DP.
1. DP upgrade activates plugins that were not active before. This is
because we were blindly calling activate on these modules. Fixed by
checking to see if the plugin was active before the update and only
calling activate afterwards as appropriate.

2. The framework was considering locked packages to be upgradeable so
we'd try and fail to upgrade those if we chose the "upgrade all"
option. Excluding locked packages from
GalleryRepository::getAllUpgradeablePackages() fixes this.