Don't let DP upgrade you to an incompatible module
This addresses two problems that can lead you to download an
incompatible module via DP. If you download an index which has a
module with incompatible core or plugin APIs, the AdminRepository page
will not show an upgrade (no upgrade icon, installed version will
match available version, and the link is "modify" not "download") but
there are two ways that you can accidentally install it:

1. Click on the "modify" link and you'll see the current version as
well as the incompatible version. If you choose the incompatible
version, it'll install it for you.

2. If you click "upgrade all" it'll pick one of the multiple plugins
available (the newest one, I think) and upgrade you to that version,
even if it's incompatible.

Fix this by comparing APIs and tracking whether or not the plugin is
compatible, then check that when we call
GalleryRepository::getAllUpgradeablePackages (this takes care of the
"upgrade all" button). In AdminRepositoryDownload, we check to see if
the plugin is compatible before showing it in the UI.

The rest of this change is test code. Because we now care about the
specific API, I've modified the test data to use the real API
versions. This required me to update tests to use the real API
versions, which looks a lot worse than it really is. And the
repository tests are these terrible data-driven tests that are painful
to work with.

For 2.2.1 I think we only want to push the non-test changes.

Review #169