Fix core upgrade notification in template
An incorrect variable name (which, sadly we don't get a warning for
from Smarty) causes us to fail to see the notification that there's a
newer version of the core available. And a quick review of the text
shows that one of the bullet points is confusing/unhelpful so I
removed it.

Review #173.