Log PHP errors into our new Event Log
We do this by setting a PHP error handler.  Unfortunately we can only
do this for PHP4 because PHP5 requires static methods to be marked
with the static keyword, or it generates an error. PHP4 cannot parse
that syntax, though so we can either break support for PHP4, or suffer
reams of E_STRICT warnings on PHP5.  We get so many of these bogus
E_STRICT warnings that it's a performance hit to funnel them through
our handler.  Andy measured this at 3-4%.  So for now, only set the
handler for PHP4.