Provide a way for localizers to copy .mo files into g2data/locale
Figure out where the .mo file should live in the g2data/locale
hierarchy and copy the newly compiled file there, whether or not the
file already existed in that location.  If the copy fails, or if we
can't find the g2data dir then warn the user that they may have to run
the filesystem permissions support task to fix up permissions.

Create a new operation in the filesystem permissions support task to
allow you to change the permissions just on the locale dir. You can do
it using the advanced box, of course, but the new operation makes it
easier for them to find.