Edge case fix and usability improvements. See review 220.
1. Fix:
When deactvating the last toolkit that provides PPM support, it still shows the "square thumb" link in the sidebar and there, one can still change from "crop" to "fit", which is actually no longer supported.

2. Usability:
Users are confused about things disapear, options that are only available under certain conditions.

Therefore, we always show the "square thumb" link from now on and don't offer the 'fit' mode if it's not available.