Fix for bug 1641893 - Handle transparency when resizing / changing file format. Review 245. Original code is from phpthumb by James Heinrich.
The first part of the fix is about preparing the destination GD image resource for transparency before we copy / blend any data to it.

The second part of the fix is about converting the alpha channel data of the resulting image resource if the output format can't store a complete alpha channel.
For JPG, we need to replace the transparent area with a background color.
For GIF (a single indexed transparent color instead of a full alpha channel), we need to switch from true color mode to an indexed color palette (dithering) and then set the pixels that are mostly transparent to the transparent color.
The process for transparent GIFs is pretty expensive since we need to scan each pixel manually twice and set each transparent pixel to the transparent color one by one as well.
For PNG and other formats with a full alpha channel, we don't need to do anything in the second pass.