Add Ajax delete comment feature.
This allows you to delete comments inline on the comment.ShowComment
and comment.ShowAllComments pages. It greatly speeds up removing spam
comments from a Gallery install.

Problem with the current code is that after you delete comments from a
page, you need to refresh the page to refill the spots that were
deleted. This is not great, but it's not trivial to have it fill those
in again, either.

Review #288: http://reviews.gallery2.org/index.jsp?page=ReviewDisplay&reviewid=288