Add rel="nofollow" to all bbcode generated anchors
This is heavy handed. Really, we only want to do it for user generated
content, and comments are obviously the first best target. However,
it's not trivial to pass through an option to have the comment code
add this (I'd like to make it something like "|markup:untrusted"). And
it's also unclear if we really trust people to add captions properly,
etc. And adding rel="nofollow" shouldn't hurt much.

Review #369