Fix for Bug 1824635 Notification Module Causes Add Item to Fail.
The problem is that GalleryEntity::save posts the save event before it has actually persisted the entity to the database.  This restricts the save handler from doing any calls that attempt to load the entity.

Changed the notification handler to pass the item object around instead of the entity id.