Clean up the Multisite step.
Remove a lot of the extra words to make the general intent clean.
Disable the Multisite text box unless you select the multisite radio
button so that users don't enter stuff unless they've selected the box.

Disabling the text box also means that we're not passing the
configPath back on every form post anymore so we need to make some
minor adjustments in the step code.  Also, I tweaked the disabled
class in the CSS to make it more obvious that you can't type stuff
into the text box until you select the radio button.