Work around shutdown issues with GalleryUrlGenerator::_getLocalPath on PHP 5.3
Our approach for registering the shutdown action doesn't work properly
in PHP 5.3. I'm not sure what goes wrong exactly, but something goes
sideways with our references and GalleryUrlGenerator::_saveLocalUrlMap()
is never even called. I'm confident that this is a bug in PHP since
the function should at least be *called*, but since it may not be
fixed before 5.3 goes gold, we can work around it by dropping the
references altogether. That approach should work for all of PHP5
anyway (but for now I just scope it back to PHP 5.3+).

Review 791:
http://reviews.gallery2.org/index.jsp?page=ReviewDisplay&reviewid=791