Make all of our urls more RESTful.  Now all of our urls are to a direct resource, upon which we can take actions.  So for example, an album is:   /album/1
To add to it, you'd GET or POST to:
  /album/1/add

Any photo is available by:
  /photo/2

And you could edit a photo's details with
  /photo/2/edit

This is going to have some interesting consequences when we want to
start showing the path to the image in the url, since eventually you
could imagine that we'd have a single resource like this:
  /Family/Videos/Bob

So this means that you couldn't actually give a resource a verb name.
But that's off in the future and this is a prototype so I'm not going
to worry about that for now.