Gut the comment module and simplify it.  Stop trying to support Atom and XML for now, we have no driver for those technologies so anything we implement is not going to be sufficiently tested and therefore it'll be broken.
Change all comment functions to return JSON and update the JS to deal
purely with JSON.  This is our new protocol for talking to the browser
and it should be flexible and portable.

Create comments.html.php.  This duplicates comment.html.php, but will
be more efficient for rendering comments since we won't be creating a
new View for every comment we render.