Say hello to the new Flash based uploader, courtesy of Eric Zelermyer.
This requires a little trickery to proxy the session id and user agent
through the ActionScript code so that we can assume the same session
in the uploader.  It's also using its own path to add photos since
we'll want to have a slightly different protocol for dealing with
responses (as opposed to JSON or HTML).

A work in progress for sure, but it's already better than what we had before.